Check if a file exists before trying to load it.
If the user attempts to load a file which is not there, show a critical
QMessageBox with a short message and the full expected filepath.  For example,
this code will run if  a user tries to load a file from the
'Open recent' menu which has been renamed or moved.