Fix two crashes in the montage workflow.
(1) When setting the montage, if the user picks a montage whose channels have
no names in common with the dataset, the application would crash. Catch this
error instead and print it out on the console.

TODO: Log this error somewhere, and consider displaying it in a QMessageBox.

(2) When plotting the current montage, calling Montage.plot() has different
behavior and a different API in version v0.14 [1] and v0.15.dev0 [2], so the
`kind` parameter will crash the application if the user installed mne from the
PyPI.

[1] https://www.martinos.org/mne/stable/generated/mne.channels.Montage.html#mne.channels.Montage.plot
[2] https://www.martinos.org/mne/dev/generated/mne.channels.Montage.html#mne.channels.Montage.plot