accomodate changes to LabRecorder nominal srate
LabRecorder was modified so that the nominal sampling rate is now a string of a double rather than a string of an int. This is actually correct because the type in the lsl::info structure is in fact a double. This means that XDF files recorded with later versions of LabRecorder (unfortunately I couldn't track down the actual commit that made this change) will throw an error at line 232. This small change should prevent this and it should remain back compatible.